fix::norm
