+++ includes/error.h
@@ -6,7 +6,7 @@
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/29 17:54:37 by eunbison          #+#    #+#             */
-/*   Updated: 2023/02/10 10:39:24 by eunson           ###   ########.fr       */
+/*   Updated: 2023/02/17 18:26:51 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,4 +20,5 @@
 #define INVALID_DATA "Invalid Data"
 #define INVALID_RANGE "Invalid Range"
 #define MALLOC_FAIL "Malloc Fail"
+
 #endif
