+++ src/calculation/degree_to_radian.c
@@ -1,26 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ray_operation.c                                    :+:      :+:    :+:   */
+/*   degree_to_radian.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/10 18:23:13 by eunson            #+#    #+#             */
-/*   Updated: 2023/02/17 21:16:59 by eunson           ###   ########.fr       */
+/*   Created: 2023/02/17 21:26:40 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 21:29:18 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_point3	ray_at(t_ray *ray, double t)
+double	degree_to_radian(double degree)
 {
-	t_point3	at;
-
-	at = v_plus(ray->orig, v_mult(ray->dir_vector, t));
-	return (at);
-}
-
-t_vector3	get_reflect(t_vector3 v, t_vector3 n)
-{
-	return (v_minus(v, v_mult(n, v_dot(v, n) * 2)));
+	return (degree * M_PI / 180.0);
 }
