+++ src/calculation/vec3_operation.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_operation.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: minsukan <minsukan@student.42.fr>          +#+  +:+       +#+        */
+/*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 23:35:32 by eunson            #+#    #+#             */
-/*   Updated: 2023/02/17 11:59:23 by minsukan         ###   ########.fr       */
+/*   Updated: 2023/02/17 16:27:33 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,20 +22,14 @@ double	v_len(t_vector3 v1)
 	return (sqrt(v_len_square(v1)));
 }
 
+double	vector_value(t_vector3 v1)
+{
+	return (v1.x + v1.y + v1.z);
+}
+
 void	set_vec3(t_vector3 *vec3, double x, double y, double z)
 {
 	vec3->x = x;
 	vec3->y = y;
 	vec3->z = z;
-}
-
-t_vector3	v_minus_(t_vector3 point)
-{
-	t_vector3	v1;
-	
-	v1.x = point.x *= -1;
-	v1.y = point.y *= -1;
-	v1.z = point.z *= -1;
-
-	return (v1);
 }
