+++ src/calculation/vec3_operation3.c
@@ -5,55 +5,33 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/07 12:16:33 by minsukan          #+#    #+#             */
-/*   Updated: 2023/02/10 10:41:14 by eunson           ###   ########.fr       */
+/*   Created: 2023/02/17 16:26:01 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 16:26:16 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-double	v_dot(t_vector3 v1, t_vector3 v2)
+t_vector3	v_mult(t_vector3 v1, double s)
 {
-	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
-}
-
-t_vector3	v_cross(t_vector3 v1, t_vector3 v2)
-{
-	t_vector3	v3;
-
-	v3.x = v1.y * v2.z - v1.z * v2.y;
-	v3.y = v1.z * v2.x - v1.x * v2.z;
-	v3.z = v1.x * v2.y - v1.y * v2.x;
-	return (v3);
-}
-
-t_vector3	v_unit(t_vector3 v1)
-{
-	double	len;
-
-	len = v_len(v1);
-	if (len == 0)
-		print_error_with_exit(ZERO_DENOMINATOR);
-	v1.x /= len;
-	v1.y /= len;
-	v1.z /= len;
+	v1.x *= s;
+	v1.y *= s;
+	v1.z *= s;
 	return (v1);
 }
 
-t_vector3	v_min(t_vector3 v1, t_vector3 v2)
+t_vector3	v_mult_(t_vector3 v1, t_vector3 v2)
 {
-	if (v1.x > v2.x)
-		v1.x = v2.x;
-	if (v1.y > v2.y)
-		v1.y = v2.y;
-	if (v1.z > v2.z)
-		v1.z = v2.z;
+	v1.x *= v2.x;
+	v1.y *= v2.y;
+	v1.z *= v2.z;
 	return (v1);
 }
 
-void	v_plus_(t_vector3 *v1, t_vector3 v2)
+t_vector3	v_divide(t_vector3 v1, double s)
 {
-	v1->x += v2.x;
-	v1->y += v2.y;
-	v1->z += v2.z;
+	v1.x *= 1 / s;
+	v1.y *= 1 / s;
+	v1.z *= 1 / s;
+	return (v1);
 }
