+++ src/constructor/viewport_construcotr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   viewport_construcotr.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: minsukan <minsukan@student.42.fr>          +#+  +:+       +#+        */
+/*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/08 10:30:24 by minsukan          #+#    #+#             */
-/*   Updated: 2023/02/16 15:00:22 by minsukan         ###   ########.fr       */
+/*   Updated: 2023/02/17 21:26:55 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,15 +22,11 @@ t_viewport	c_viewport(t_camera *camera)
 	viewport.h_vector = c_vector3(viewport.width, 0, 0);
 	viewport.v_vector = c_vector3(0, viewport.height, 0);
 	viewport.left_bottom = \
-			v_minus(v_minus(v_minus(camera->point, v_divide(viewport.h_vector, 2)), \
+			v_minus(v_minus(v_minus(camera->point, \
+			v_divide(viewport.h_vector, 2)), \
 			v_divide(viewport.v_vector, 2)), \
 			c_vector3(0, 0, viewport.focal_len));
 	return (viewport);
-}
-
-double	degress_to_radians(double degrees)
-{
-	return (degrees * M_PI / 180.0);
 }
 
 t_viewport update_viewport(t_camera *camera)
