+++ src/draw/draw_image.c
@@ -1,26 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ray_operation.c                                    :+:      :+:    :+:   */
+/*   draw_image.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/10 18:23:13 by eunson            #+#    #+#             */
-/*   Updated: 2023/02/17 21:16:59 by eunson           ###   ########.fr       */
+/*   Created: 2023/02/17 16:46:01 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 16:46:17 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_point3	ray_at(t_ray *ray, double t)
+void	draw_image(t_info *info)
 {
-	t_point3	at;
-
-	at = v_plus(ray->orig, v_mult(ray->dir_vector, t));
-	return (at);
-}
-
-t_vector3	get_reflect(t_vector3 v, t_vector3 n)
-{
-	return (v_minus(v, v_mult(n, v_dot(v, n) * 2)));
+	mlx_clear_window(info->mlx_info.mlx_ptr, info->mlx_info.win_ptr);
+	info->scene.camera->viewport = update_viewport(info->scene.camera);
+	draw_scene(info);
 }
