+++ src/draw/draw_scene.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   draw_scene.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: minsukan <minsukan@student.42.fr>          +#+  +:+       +#+        */
+/*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/08 11:57:35 by minsukan          #+#    #+#             */
-/*   Updated: 2023/02/17 13:53:15 by minsukan         ###   ########.fr       */
+/*   Updated: 2023/02/17 16:17:01 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-void	img_pix_put(t_image *img, int x, int y, int color)
+static void	img_pix_put(t_image *img, int x, int y, int color)
 {
 	char	*pixel;
 
@@ -20,13 +20,10 @@ void	img_pix_put(t_image *img, int x, int y, int color)
 	*(int *)pixel = color;
 }
 
-int	convert_rgb(t_rgb rgb)
+static void	set_camera(t_info *info)
 {
-	int	color; 
-
-	color = 0;
-	color = ((int)(255.0 * rgb.x) << 16) + ((int)(255.0 * rgb.y) << 8) + ((int)(255.0 * rgb.z));
-	return (color);
+	move_scene(info, v_minus_(info->scene.camera->point));
+	info->scene.camera->point = c_vector3(0,0,0);
 }
 
 void	draw_scene(t_info *info)
@@ -38,6 +35,7 @@ void	draw_scene(t_info *info)
 	double	v;
 
 	//t_rgb	color;
+	set_camera(info);
 	height = IMG_HEIGHT - 1;
 	while (height >= 0)
 	{
@@ -61,7 +59,5 @@ void	draw_scene(t_info *info)
 		}
 		height--;
 	}
-	printf("camera %f %f %f\n", info->scene.camera->point.x, info->scene.camera->point.y, info->scene.camera->point.z);
-	//printf("dir x : %f, y : %f, z : %f\n", info->scene.camera->dir_vector.x, info->scene.camera->dir_vector.y, info->scene.camera->dir_vector.z);
 	mlx_put_image_to_window(info->mlx_info.mlx_ptr, info->mlx_info.win_ptr, info->mlx_info.img_ptr, 0, 0);
 }
