+++ src/event/move_scene.c
@@ -1,26 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ray_operation.c                                    :+:      :+:    :+:   */
+/*   move_scene.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/10 18:23:13 by eunson            #+#    #+#             */
-/*   Updated: 2023/02/17 21:16:59 by eunson           ###   ########.fr       */
+/*   Created: 2023/02/17 16:53:39 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 16:53:53 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_point3	ray_at(t_ray *ray, double t)
+void	move_scene(t_info *info, t_vector3 dir)
 {
-	t_point3	at;
-
-	at = v_plus(ray->orig, v_mult(ray->dir_vector, t));
-	return (at);
-}
-
-t_vector3	get_reflect(t_vector3 v, t_vector3 n)
-{
-	return (v_minus(v, v_mult(n, v_dot(v, n) * 2)));
+	move_figure(info->scene.figures, dir);
+	move_light(info->scene.lights, dir);
 }
