+++ src/event/option_control.c
@@ -1,26 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ray_operation.c                                    :+:      :+:    :+:   */
+/*   option_control.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/10 18:23:13 by eunson            #+#    #+#             */
-/*   Updated: 2023/02/17 21:16:59 by eunson           ###   ########.fr       */
+/*   Created: 2023/02/17 16:46:52 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 16:47:03 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_point3	ray_at(t_ray *ray, double t)
+void	option_control(int keycode, t_info *info)
 {
-	t_point3	at;
-
-	at = v_plus(ray->orig, v_mult(ray->dir_vector, t));
-	return (at);
-}
-
-t_vector3	get_reflect(t_vector3 v, t_vector3 n)
-{
-	return (v_minus(v, v_mult(n, v_dot(v, n) * 2)));
+	if (keycode == KEY_1)
+	{
+		if (info->scene.option.shadow == True)
+			info->scene.option.shadow = False;
+		else
+			info->scene.option.shadow = True;
+	}
 }
