+++ src/event/rotate_light.c
@@ -1,0 +1,31 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   rotate_light.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/02/17 16:42:17 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 21:34:16 by eunson           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minirt.h"
+
+static void	get_light_point(t_light *light, \
+							t_vector3 dir_x, t_vector3 dir_y, t_vector3 dir_z)
+{
+	light->point = c_vector3(vector_value(v_mult_(dir_x, light->point)),
+			vector_value(v_mult_(dir_y, light->point)),
+			vector_value(v_mult_(dir_z, light->point)));
+}
+
+void	rotate_light(t_list *light, \
+							t_vector3 dir_x, t_vector3 dir_y, t_vector3 dir_z)
+{
+	while (light)
+	{
+		get_light_point((t_light *)light->obj, dir_x, dir_y, dir_z);
+		light = light->next;
+	}
+}
