+++ src/event/update_fov.c
@@ -1,26 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ray_operation.c                                    :+:      :+:    :+:   */
+/*   update_fov.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: eunson <eunson@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/10 18:23:13 by eunson            #+#    #+#             */
-/*   Updated: 2023/02/17 21:16:59 by eunson           ###   ########.fr       */
+/*   Created: 2023/02/17 16:46:33 by eunson            #+#    #+#             */
+/*   Updated: 2023/02/17 16:46:44 by eunson           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_point3	ray_at(t_ray *ray, double t)
+void	update_fov(int keycode, t_info *info)
 {
-	t_point3	at;
-
-	at = v_plus(ray->orig, v_mult(ray->dir_vector, t));
-	return (at);
-}
-
-t_vector3	get_reflect(t_vector3 v, t_vector3 n)
-{
-	return (v_minus(v, v_mult(n, v_dot(v, n) * 2)));
+	if (keycode == KEY_UP && info->scene.camera->fov < 180)
+		info->scene.camera->fov++;
+	if (keycode == KEY_DOWN && info->scene.camera->fov > 0)
+		info->scene.camera->fov--;
 }
